requirements.txt
alembic==1.4.2
amqp==2.6.1
billiard==3.6.3.0
cachelib==0.1.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
dnspython==1.16.0
eventlet==0.26.1
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==0.4.16
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
Mako==1.1.3
MarkupSafe==1.1.1
monotonic==1.5
Pillow==7.2.0
pycparser==2.20
pycryptodomex==3.9.4
PyMySQL==0.10.0
pyOpenSSL==19.1.0
python-alipay-sdk==2.1.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
qiniu==7.2.10
redis==3.5.3
requests==2.24.0
ronglian-sms-sdk==1.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
vine==1.3.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.1.0
